Address review comments on cmd package
